requirements.txt
tensorflow==2.15.0
gradio
pillow
numpy<2.0.0